+TelaPacienteVisualizacao, +itemPaciente, ++TelaMedico

+++ app/src/main/java/com/projeto/healthmonitor/model/RegistroDiario.kt
@@ -11,6 +11,6 @@ data class RegistroDiario(
     val id: Long = 0,
     val pacienteId: Long,
     val data: String,
-    val pressaoSistolica: Int,
+    val pressaoSistolica: Float,
     val glicemia: Float
 )
